requirements.txt
biopython==1.83
ipython==8.12.3
model==0.6.0
numpy==2.2.4
pandas==2.2.3
scikit_learn==1.4.2
tensorflow==2.17.0
torch==1.11.0+cu115
tqdm==4.66.2
transformers==4.28.0